Organize provider settings into separate components (#3196)
