+++ .changeset/five-pigs-talk.md
@@ -1,0 +1,5 @@
+---
+"roo-cline": patch
+---
+
+Organize provider settings into separate components
