+++ webview-ui/src/components/settings/ApiOptions.tsx
@@ -1,31 +1,17 @@
 import React, { memo, useCallback, useEffect, useMemo, useState } from "react"
-import { useDebounce, useEvent } from "react-use"
-import { Trans } from "react-i18next"
-import { LanguageModelChatSelector } from "vscode"
-import { Checkbox } from "vscrui"
-import {
-	VSCodeButton,
-	VSCodeLink,
-	VSCodeRadio,
-	VSCodeRadioGroup,
-	VSCodeTextField,
-} from "@vscode/webview-ui-toolkit/react"
+import { useDebounce } from "react-use"
+import { VSCodeLink, VSCodeTextField } from "@vscode/webview-ui-toolkit/react"
 import { ExternalLinkIcon } from "@radix-ui/react-icons"
 
-import { ReasoningEffort as ReasoningEffortType } from "@roo/schemas"
 import {
 	ApiConfiguration,
-	ModelInfo,
-	azureOpenAiDefaultApiVersion,
 	glamaDefaultModelId,
 	mistralDefaultModelId,
-	openAiModelInfoSaneDefaults,
 	openRouterDefaultModelId,
 	unboundDefaultModelId,
 	requestyDefaultModelId,
 	ApiProvider,
 } from "@roo/shared/api"
-import { ExtensionMessage } from "@roo/shared/ExtensionMessage"
 
 import { vscode } from "@src/utils/vscode"
 import { validateApiConfiguration, validateModelId, validateBedrockArn } from "@src/utils/validate"
@@ -36,18 +22,28 @@ import {
 	useOpenRouterModelProviders,
 	OPENROUTER_DEFAULT_PROVIDER_NAME,
 } from "@src/components/ui/hooks/useOpenRouterModelProviders"
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue, Button } from "@src/components/ui"
-import { getRequestyAuthUrl, getOpenRouterAuthUrl, getGlamaAuthUrl } from "@src/oauth/urls"
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@src/components/ui"
+import { VSCodeButtonLink } from "@src/components/common/VSCodeButtonLink"
+import { getRequestyAuthUrl, getGlamaAuthUrl } from "@src/oauth/urls"
 
-import { VSCodeButtonLink } from "../common/VSCodeButtonLink"
+// Providers
+import { Anthropic } from "./providers/Anthropic"
+import { Bedrock } from "./providers/Bedrock"
+import { Gemini } from "./providers/Gemini"
+import { LMStudio } from "./providers/LMStudio"
+import { Ollama } from "./providers/Ollama"
+import { OpenAI } from "./providers/OpenAI"
+import { OpenAICompatible } from "./providers/OpenAICompatible"
+import { OpenRouter } from "./providers/OpenRouter"
+import { Vertex } from "./providers/Vertex"
+import { VSCodeLM } from "./providers/VSCodeLM"
 
-import { MODELS_BY_PROVIDER, PROVIDERS, VERTEX_REGIONS, REASONING_MODELS, AWS_REGIONS } from "./constants"
+import { MODELS_BY_PROVIDER, PROVIDERS, REASONING_MODELS } from "./constants"
+import { inputEventTransform, noTransform } from "./transforms"
 import { ModelInfoView } from "./ModelInfoView"
 import { ModelPicker } from "./ModelPicker"
 import { ApiErrorMessage } from "./ApiErrorMessage"
 import { ThinkingBudget } from "./ThinkingBudget"
-import { R1FormatSetting } from "./R1FormatSetting"
-import { OpenRouterBalanceDisplay } from "./OpenRouterBalanceDisplay"
 import { RequestyBalanceDisplay } from "./RequestyBalanceDisplay"
 import { ReasoningEffort } from "./ReasoningEffort"
 import { PromptCachingControl } from "./PromptCachingControl"
@@ -74,12 +70,6 @@ const ApiOptions = ({
 }: ApiOptionsProps) => {
 	const { t } = useAppTranslation()
 
-	const [ollamaModels, setOllamaModels] = useState<string[]>([])
-	const [lmStudioModels, setLmStudioModels] = useState<string[]>([])
-	const [vsCodeLmModels, setVsCodeLmModels] = useState<LanguageModelChatSelector[]>([])
-
-	const [openAiModels, setOpenAiModels] = useState<Record<string, ModelInfo> | null>(null)
-
 	const [customHeaders, setCustomHeaders] = useState<[string, string][]>(() => {
 		const headers = apiConfiguration?.openAiHeaders || {}
 		return Object.entries(headers)
@@ -93,67 +83,28 @@ const ApiOptions = ({
 		}
 	}, [apiConfiguration?.openAiHeaders, customHeaders])
 
-	const [anthropicBaseUrlSelected, setAnthropicBaseUrlSelected] = useState(!!apiConfiguration?.anthropicBaseUrl)
-	const [openAiNativeBaseUrlSelected, setOpenAiNativeBaseUrlSelected] = useState(
-		!!apiConfiguration?.openAiNativeBaseUrl,
-	)
-	const [azureApiVersionSelected, setAzureApiVersionSelected] = useState(!!apiConfiguration?.azureApiVersion)
-	const [openRouterBaseUrlSelected, setOpenRouterBaseUrlSelected] = useState(!!apiConfiguration?.openRouterBaseUrl)
-	const [openAiLegacyFormatSelected, setOpenAiLegacyFormatSelected] = useState(!!apiConfiguration?.openAiLegacyFormat)
-	const [googleGeminiBaseUrlSelected, setGoogleGeminiBaseUrlSelected] = useState(
-		!!apiConfiguration?.googleGeminiBaseUrl,
-	)
-
-	const handleAddCustomHeader = useCallback(() => {
-		// Only update the local state to show the new row in the UI
-		setCustomHeaders((prev) => [...prev, ["", ""]])
-		// Do not update the main configuration yet, wait for user input
-	}, [])
-
-	const handleUpdateHeaderKey = useCallback((index: number, newKey: string) => {
-		setCustomHeaders((prev) => {
-			const updated = [...prev]
-			if (updated[index]) {
-				updated[index] = [newKey, updated[index][1]]
-			}
-			return updated
-		})
-	}, [])
-
-	const handleUpdateHeaderValue = useCallback((index: number, newValue: string) => {
-		setCustomHeaders((prev) => {
-			const updated = [...prev]
-			if (updated[index]) {
-				updated[index] = [updated[index][0], newValue]
-			}
-			return updated
-		})
-	}, [])
-
-	const handleRemoveCustomHeader = useCallback((index: number) => {
-		setCustomHeaders((prev) => prev.filter((_, i) => i !== index))
-	}, [])
-
-	// Helper to convert array of tuples to object (filtering out empty keys)
+	// Helper to convert array of tuples to object (filtering out empty keys).
 	const convertHeadersToObject = (headers: [string, string][]): Record<string, string> => {
 		const result: Record<string, string> = {}
 
-		// Process each header tuple
+		// Process each header tuple.
 		for (const [key, value] of headers) {
 			const trimmedKey = key.trim()
 
-			// Skip empty keys
-			if (!trimmedKey) continue
+			// Skip empty keys.
+			if (!trimmedKey) {
+				continue
+			}
 
-			// For duplicates, the last one in the array wins
-			// This matches how HTTP headers work in general
+			// For duplicates, the last one in the array wins.
+			// This matches how HTTP headers work in general.
 			result[trimmedKey] = value.trim()
 		}
 
 		return result
 	}
 
-	// Debounced effect to update the main configuration when local customHeaders state stabilizes
+	// Debounced effect to update the main configuration when local customHeaders state stabilizes.
 	useDebounce(
 		() => {
 			const currentConfigHeaders = apiConfiguration?.openAiHeaders || {}
@@ -169,9 +120,6 @@ const ApiOptions = ({
 	)
 
 	const [isDescriptionExpanded, setIsDescriptionExpanded] = useState(false)
-	const noTransform = <T,>(value: T) => value
-
-	const inputEventTransform = <E,>(event: E) => (event as { target: HTMLInputElement })?.target?.value as any
 
 	const handleInputChange = useCallback(
 		<K extends keyof ApiConfiguration, E>(
@@ -204,8 +152,9 @@ const ApiOptions = ({
 	useDebounce(
 		() => {
 			if (selectedProvider === "openai") {
-				// Use our custom headers state to build the headers object
+				// Use our custom headers state to build the headers object.
 				const headerObject = convertHeadersToObject(customHeaders)
+
 				vscode.postMessage({
 					type: "requestOpenAiModels",
 					values: {
@@ -238,6 +187,7 @@ const ApiOptions = ({
 	useEffect(() => {
 		const apiValidationResult =
 			validateApiConfiguration(apiConfiguration) || validateModelId(apiConfiguration, routerModels)
+
 		setErrorMessage(apiValidationResult)
 	}, [apiConfiguration, routerModels, setErrorMessage])
 
@@ -250,38 +200,6 @@ const ApiOptions = ({
 			apiConfiguration.openRouterModelId in routerModels.openrouter,
 	})
 
-	const onMessage = useCallback((event: MessageEvent) => {
-		const message: ExtensionMessage = event.data
-
-		switch (message.type) {
-			case "openAiModels": {
-				const updatedModels = message.openAiModels ?? []
-				setOpenAiModels(Object.fromEntries(updatedModels.map((item) => [item, openAiModelInfoSaneDefaults])))
-				break
-			}
-			case "ollamaModels":
-				{
-					const newModels = message.ollamaModels ?? []
-					setOllamaModels(newModels)
-				}
-				break
-			case "lmStudioModels":
-				{
-					const newModels = message.lmStudioModels ?? []
-					setLmStudioModels(newModels)
-				}
-				break
-			case "vsCodeLmModels":
-				{
-					const newModels = message.vsCodeLmModels ?? []
-					setVsCodeLmModels(newModels)
-				}
-				break
-		}
-	}, [])
-
-	useEvent("message", onMessage)
-
 	const selectedProviderModelOptions = useMemo(
 		() =>
 			MODELS_BY_PROVIDER[selectedProvider]
@@ -293,16 +211,13 @@ const ApiOptions = ({
 		[selectedProvider],
 	)
 
-	// Base URL for provider documentation
-	const DOC_BASE_URL = "https://docs.roocode.com/providers"
-
-	// Custom URL path mappings for providers with different slugs
+	// Custom URL path mappings for providers with different slugs.
 	const providerUrlSlugs: Record<string, string> = {
 		"openai-native": "openai",
 		openai: "openai-compatible",
 	}
 
-	// Helper function to get provider display name from PROVIDERS constant
+	// Helper function to get provider display name from PROVIDERS constant.
 	const getProviderDisplayName = (providerKey: string): string | undefined => {
 		const provider = PROVIDERS.find((p) => p.value === providerKey)
 		return provider?.label
@@ -320,7 +235,7 @@ const ApiOptions = ({
 		const urlSlug = providerUrlSlugs[selectedProvider] || selectedProvider
 
 		return {
-			url: `${DOC_BASE_URL}/${urlSlug}`,
+			url: `https://docs.roocode.com/providers/${urlSlug}`,
 			name: displayName,
 		}
 	}
@@ -403,126 +318,16 @@ const ApiOptions = ({
 			{errorMessage && <ApiErrorMessage errorMessage={errorMessage} />}
 
 			{selectedProvider === "openrouter" && (
-				<>
-					<VSCodeTextField
-						value={apiConfiguration?.openRouterApiKey || ""}
-						type="password"
-						onInput={handleInputChange("openRouterApiKey")}
-						placeholder={t("settings:placeholders.apiKey")}
-						className="w-full">
-						<div className="flex justify-between items-center mb-1">
-							<label className="block font-medium">{t("settings:providers.openRouterApiKey")}</label>
-							{apiConfiguration?.openRouterApiKey && (
-								<OpenRouterBalanceDisplay
-									apiKey={apiConfiguration.openRouterApiKey}
-									baseUrl={apiConfiguration.openRouterBaseUrl}
-								/>
-							)}
-						</div>
-					</VSCodeTextField>
-					<div className="text-sm text-vscode-descriptionForeground -mt-2">
-						{t("settings:providers.apiKeyStorageNotice")}
-					</div>
-					{!apiConfiguration?.openRouterApiKey && (
-						<VSCodeButtonLink
-							href={getOpenRouterAuthUrl(uriScheme)}
-							style={{ width: "100%" }}
-							appearance="primary">
-							{t("settings:providers.getOpenRouterApiKey")}
-						</VSCodeButtonLink>
-					)}
-					{!fromWelcomeView && (
-						<>
-							<div>
-								<Checkbox
-									checked={openRouterBaseUrlSelected}
-									onChange={(checked: boolean) => {
-										setOpenRouterBaseUrlSelected(checked)
-
-										if (!checked) {
-											setApiConfigurationField("openRouterBaseUrl", "")
-										}
-									}}>
-									{t("settings:providers.useCustomBaseUrl")}
-								</Checkbox>
-								{openRouterBaseUrlSelected && (
-									<VSCodeTextField
-										value={apiConfiguration?.openRouterBaseUrl || ""}
-										type="url"
-										onInput={handleInputChange("openRouterBaseUrl")}
-										placeholder="Default: https://openrouter.ai/api/v1"
-										className="w-full mt-1"
-									/>
-								)}
-							</div>
-							<Checkbox
-								checked={apiConfiguration?.openRouterUseMiddleOutTransform ?? true}
-								onChange={handleInputChange("openRouterUseMiddleOutTransform", noTransform)}>
-								<Trans
-									i18nKey="settings:providers.openRouterTransformsText"
-									components={{
-										// eslint-disable-next-line jsx-a11y/anchor-has-content
-										a: <a href="https://openrouter.ai/docs/transforms" />,
-									}}
-								/>
-							</Checkbox>
-						</>
-					)}
-				</>
+				<OpenRouter
+					apiConfiguration={apiConfiguration}
+					setApiConfigurationField={setApiConfigurationField}
+					uriScheme={uriScheme}
+					fromWelcomeView={fromWelcomeView}
+				/>
 			)}
 
 			{selectedProvider === "anthropic" && (
-				<>
-					<VSCodeTextField
-						value={apiConfiguration?.apiKey || ""}
-						type="password"
-						onInput={handleInputChange("apiKey")}
-						placeholder={t("settings:placeholders.apiKey")}
-						className="w-full">
-						<label className="block font-medium mb-1">{t("settings:providers.anthropicApiKey")}</label>
-					</VSCodeTextField>
-					<div className="text-sm text-vscode-descriptionForeground -mt-2">
-						{t("settings:providers.apiKeyStorageNotice")}
-					</div>
-					{!apiConfiguration?.apiKey && (
-						<VSCodeButtonLink href="https://console.anthropic.com/settings/keys" appearance="secondary">
-							{t("settings:providers.getAnthropicApiKey")}
-						</VSCodeButtonLink>
-					)}
-					<div>
-						<Checkbox
-							checked={anthropicBaseUrlSelected}
-							onChange={(checked: boolean) => {
-								setAnthropicBaseUrlSelected(checked)
-
-								if (!checked) {
-									setApiConfigurationField("anthropicBaseUrl", "")
-									setApiConfigurationField("anthropicUseAuthToken", false) // added
-								}
-							}}>
-							{t("settings:providers.useCustomBaseUrl")}
-						</Checkbox>
-						{anthropicBaseUrlSelected && (
-							<>
-								<VSCodeTextField
-									value={apiConfiguration?.anthropicBaseUrl || ""}
-									type="url"
-									onInput={handleInputChange("anthropicBaseUrl")}
-									placeholder="https://api.anthropic.com"
-									className="w-full mt-1"
-								/>
-
-								{/* added */}
-								<Checkbox
-									checked={apiConfiguration?.anthropicUseAuthToken ?? false}
-									onChange={handleInputChange("anthropicUseAuthToken", noTransform)}
-									className="w-full mt-1">
-									{t("settings:providers.anthropicUseAuthToken")}
-								</Checkbox>
-							</>
-						)}
-					</div>
-				</>
+				<Anthropic apiConfiguration={apiConfiguration} setApiConfigurationField={setApiConfigurationField} />
 			)}
 
 			{selectedProvider === "glama" && (
@@ -579,46 +384,7 @@ const ApiOptions = ({
 			)}
 
 			{selectedProvider === "openai-native" && (
-				<>
-					<Checkbox
-						checked={openAiNativeBaseUrlSelected}
-						onChange={(checked: boolean) => {
-							setOpenAiNativeBaseUrlSelected(checked)
-
-							if (!checked) {
-								setApiConfigurationField("openAiNativeBaseUrl", "")
-							}
-						}}>
-						{t("settings:providers.useCustomBaseUrl")}
-					</Checkbox>
-					{openAiNativeBaseUrlSelected && (
-						<>
-							<VSCodeTextField
-								value={apiConfiguration?.openAiNativeBaseUrl || ""}
-								type="url"
-								onInput={handleInputChange("openAiNativeBaseUrl")}
-								placeholder="https://api.openai.com/v1"
-								className="w-full mt-1"
-							/>
-						</>
-					)}
-					<VSCodeTextField
-						value={apiConfiguration?.openAiNativeApiKey || ""}
-						type="password"
-						onInput={handleInputChange("openAiNativeApiKey")}
-						placeholder={t("settings:placeholders.apiKey")}
-						className="w-full">
-						<label className="block font-medium mb-1">{t("settings:providers.openAiApiKey")}</label>
-					</VSCodeTextField>
-					<div className="text-sm text-vscode-descriptionForeground -mt-2">
-						{t("settings:providers.apiKeyStorageNotice")}
-					</div>
-					{!apiConfiguration?.openAiNativeApiKey && (
-						<VSCodeButtonLink href="https://platform.openai.com/api-keys" appearance="secondary">
-							{t("settings:providers.getOpenAiApiKey")}
-						</VSCodeButtonLink>
-					)}
-				</>
+				<OpenAI apiConfiguration={apiConfiguration} setApiConfigurationField={setApiConfigurationField} />
 			)}
 
 			{selectedProvider === "mistral" && (
@@ -661,832 +427,30 @@ const ApiOptions = ({
 			)}
 
 			{selectedProvider === "bedrock" && (
-				<>
-					<VSCodeRadioGroup
-						value={apiConfiguration?.awsUseProfile ? "profile" : "credentials"}
-						onChange={handleInputChange(
-							"awsUseProfile",
-							(e) => (e.target as HTMLInputElement).value === "profile",
-						)}>
-						<VSCodeRadio value="credentials">{t("settings:providers.awsCredentials")}</VSCodeRadio>
-						<VSCodeRadio value="profile">{t("settings:providers.awsProfile")}</VSCodeRadio>
-					</VSCodeRadioGroup>
-					<div className="text-sm text-vscode-descriptionForeground -mt-3">
-						{t("settings:providers.apiKeyStorageNotice")}
-					</div>
-					{apiConfiguration?.awsUseProfile ? (
-						<VSCodeTextField
-							value={apiConfiguration?.awsProfile || ""}
-							onInput={handleInputChange("awsProfile")}
-							placeholder={t("settings:placeholders.profileName")}
-							className="w-full">
-							<label className="block font-medium mb-1">{t("settings:providers.awsProfileName")}</label>
-						</VSCodeTextField>
-					) : (
-						<>
-							<VSCodeTextField
-								value={apiConfiguration?.awsAccessKey || ""}
-								type="password"
-								onInput={handleInputChange("awsAccessKey")}
-								placeholder={t("settings:placeholders.accessKey")}
-								className="w-full">
-								<label className="block font-medium mb-1">{t("settings:providers.awsAccessKey")}</label>
-							</VSCodeTextField>
-							<VSCodeTextField
-								value={apiConfiguration?.awsSecretKey || ""}
-								type="password"
-								onInput={handleInputChange("awsSecretKey")}
-								placeholder={t("settings:placeholders.secretKey")}
-								className="w-full">
-								<label className="block font-medium mb-1">{t("settings:providers.awsSecretKey")}</label>
-							</VSCodeTextField>
-							<VSCodeTextField
-								value={apiConfiguration?.awsSessionToken || ""}
-								type="password"
-								onInput={handleInputChange("awsSessionToken")}
-								placeholder={t("settings:placeholders.sessionToken")}
-								className="w-full">
-								<label className="block font-medium mb-1">
-									{t("settings:providers.awsSessionToken")}
-								</label>
-							</VSCodeTextField>
-						</>
-					)}
-					<div>
-						<label className="block font-medium mb-1">{t("settings:providers.awsRegion")}</label>
-						<Select
-							value={apiConfiguration?.awsRegion || ""}
-							onValueChange={(value) => setApiConfigurationField("awsRegion", value)}>
-							<SelectTrigger className="w-full">
-								<SelectValue placeholder={t("settings:common.select")} />
-							</SelectTrigger>
-							<SelectContent>
-								{AWS_REGIONS.map(({ value, label }) => (
-									<SelectItem key={value} value={value}>
-										{label}
-									</SelectItem>
-								))}
-							</SelectContent>
-						</Select>
-					</div>
-					<Checkbox
-						checked={apiConfiguration?.awsUseCrossRegionInference || false}
-						onChange={handleInputChange("awsUseCrossRegionInference", noTransform)}>
-						{t("settings:providers.awsCrossRegion")}
-					</Checkbox>
-					{selectedModelInfo?.supportsPromptCache && (
-						<Checkbox
-							checked={apiConfiguration?.awsUsePromptCache || false}
-							onChange={handleInputChange("awsUsePromptCache", noTransform)}>
-							<div className="flex items-center gap-1">
-								<span>{t("settings:providers.enablePromptCaching")}</span>
-								<i
-									className="codicon codicon-info text-vscode-descriptionForeground"
-									title={t("settings:providers.enablePromptCachingTitle")}
-									style={{ fontSize: "12px" }}
-								/>
-							</div>
-						</Checkbox>
-					)}
-					<div>
-						<div className="text-sm text-vscode-descriptionForeground ml-6 mt-1">
-							{t("settings:providers.cacheUsageNote")}
-						</div>
-					</div>
-				</>
+				<Bedrock
+					apiConfiguration={apiConfiguration}
+					setApiConfigurationField={setApiConfigurationField}
+					selectedModelInfo={selectedModelInfo}
+				/>
 			)}
 
 			{selectedProvider === "vertex" && (
-				<>
-					<div className="text-sm text-vscode-descriptionForeground">
-						<div>{t("settings:providers.googleCloudSetup.title")}</div>
-						<div>
-							<VSCodeLink
-								href="https://cloud.google.com/vertex-ai/generative-ai/docs/partner-models/use-claude#before_you_begin"
-								className="text-sm">
-								{t("settings:providers.googleCloudSetup.step1")}
-							</VSCodeLink>
-						</div>
-						<div>
-							<VSCodeLink
-								href="https://cloud.google.com/docs/authentication/provide-credentials-adc#google-idp"
-								className="text-sm">
-								{t("settings:providers.googleCloudSetup.step2")}
-							</VSCodeLink>
-						</div>
-						<div>
-							<VSCodeLink
-								href="https://developers.google.com/workspace/guides/create-credentials?hl=en#service-account"
-								className="text-sm">
-								{t("settings:providers.googleCloudSetup.step3")}
-							</VSCodeLink>
-						</div>
-					</div>
-					<VSCodeTextField
-						value={apiConfiguration?.vertexJsonCredentials || ""}
-						onInput={handleInputChange("vertexJsonCredentials")}
-						placeholder={t("settings:placeholders.credentialsJson")}
-						className="w-full">
-						<label className="block font-medium mb-1">
-							{t("settings:providers.googleCloudCredentials")}
-						</label>
-					</VSCodeTextField>
-					<VSCodeTextField
-						value={apiConfiguration?.vertexKeyFile || ""}
-						onInput={handleInputChange("vertexKeyFile")}
-						placeholder={t("settings:placeholders.keyFilePath")}
-						className="w-full">
-						<label className="block font-medium mb-1">{t("settings:providers.googleCloudKeyFile")}</label>
-					</VSCodeTextField>
-					<VSCodeTextField
-						value={apiConfiguration?.vertexProjectId || ""}
-						onInput={handleInputChange("vertexProjectId")}
-						placeholder={t("settings:placeholders.projectId")}
-						className="w-full">
-						<label className="block font-medium mb-1">{t("settings:providers.googleCloudProjectId")}</label>
-					</VSCodeTextField>
-					<div>
-						<label className="block font-medium mb-1">{t("settings:providers.googleCloudRegion")}</label>
-						<Select
-							value={apiConfiguration?.vertexRegion || ""}
-							onValueChange={(value) => setApiConfigurationField("vertexRegion", value)}>
-							<SelectTrigger className="w-full">
-								<SelectValue placeholder={t("settings:common.select")} />
-							</SelectTrigger>
-							<SelectContent>
-								{VERTEX_REGIONS.map(({ value, label }) => (
-									<SelectItem key={value} value={value}>
-										{label}
-									</SelectItem>
-								))}
-							</SelectContent>
-						</Select>
-					</div>
-				</>
+				<Vertex apiConfiguration={apiConfiguration} setApiConfigurationField={setApiConfigurationField} />
 			)}
 
 			{selectedProvider === "gemini" && (
-				<>
-					<VSCodeTextField
-						value={apiConfiguration?.geminiApiKey || ""}
-						type="password"
-						onInput={handleInputChange("geminiApiKey")}
-						placeholder={t("settings:placeholders.apiKey")}
-						className="w-full">
-						<label className="block font-medium mb-1">{t("settings:providers.geminiApiKey")}</label>
-					</VSCodeTextField>
-					<div className="text-sm text-vscode-descriptionForeground -mt-2">
-						{t("settings:providers.apiKeyStorageNotice")}
-					</div>
-					{!apiConfiguration?.geminiApiKey && (
-						<VSCodeButtonLink href="https://ai.google.dev/" appearance="secondary">
-							{t("settings:providers.getGeminiApiKey")}
-						</VSCodeButtonLink>
-					)}
-					<div>
-						<Checkbox
-							checked={googleGeminiBaseUrlSelected}
-							onChange={(checked: boolean) => {
-								setGoogleGeminiBaseUrlSelected(checked)
-
-								if (!checked) {
-									setApiConfigurationField("googleGeminiBaseUrl", "")
-								}
-							}}>
-							{t("settings:providers.useCustomBaseUrl")}
-						</Checkbox>
-						{googleGeminiBaseUrlSelected && (
-							<VSCodeTextField
-								value={apiConfiguration?.googleGeminiBaseUrl || ""}
-								type="url"
-								onInput={handleInputChange("googleGeminiBaseUrl")}
-								placeholder={t("settings:defaults.geminiUrl")}
-								className="w-full mt-1"
-							/>
-						)}
-					</div>
-				</>
+				<Gemini apiConfiguration={apiConfiguration} setApiConfigurationField={setApiConfigurationField} />
 			)}
 
 			{selectedProvider === "openai" && (
-				<>
-					<VSCodeTextField
-						value={apiConfiguration?.openAiBaseUrl || ""}
-						type="url"
-						onInput={handleInputChange("openAiBaseUrl")}
-						placeholder={t("settings:placeholders.baseUrl")}
-						className="w-full">
-						<label className="block font-medium mb-1">{t("settings:providers.openAiBaseUrl")}</label>
-					</VSCodeTextField>
-					<VSCodeTextField
-						value={apiConfiguration?.openAiApiKey || ""}
-						type="password"
-						onInput={handleInputChange("openAiApiKey")}
-						placeholder={t("settings:placeholders.apiKey")}
-						className="w-full">
-						<label className="block font-medium mb-1">{t("settings:providers.openAiApiKey")}</label>
-					</VSCodeTextField>
-					<ModelPicker
-						apiConfiguration={apiConfiguration}
-						setApiConfigurationField={setApiConfigurationField}
-						defaultModelId="gpt-4o"
-						models={openAiModels}
-						modelIdKey="openAiModelId"
-						serviceName="OpenAI"
-						serviceUrl="https://platform.openai.com"
-					/>
-					<R1FormatSetting
-						onChange={handleInputChange("openAiR1FormatEnabled", noTransform)}
-						openAiR1FormatEnabled={apiConfiguration?.openAiR1FormatEnabled ?? false}
-					/>
-					<div>
-						<Checkbox
-							checked={openAiLegacyFormatSelected}
-							onChange={(checked: boolean) => {
-								setOpenAiLegacyFormatSelected(checked)
-								setApiConfigurationField("openAiLegacyFormat", checked)
-							}}>
-							{t("settings:providers.useLegacyFormat")}
-						</Checkbox>
-					</div>
-					<Checkbox
-						checked={apiConfiguration?.openAiStreamingEnabled ?? true}
-						onChange={handleInputChange("openAiStreamingEnabled", noTransform)}>
-						{t("settings:modelInfo.enableStreaming")}
-					</Checkbox>
-					<Checkbox
-						checked={apiConfiguration?.openAiUseAzure ?? false}
-						onChange={handleInputChange("openAiUseAzure", noTransform)}>
-						{t("settings:modelInfo.useAzure")}
-					</Checkbox>
-					<div>
-						<Checkbox
-							checked={azureApiVersionSelected}
-							onChange={(checked: boolean) => {
-								setAzureApiVersionSelected(checked)
-
-								if (!checked) {
-									setApiConfigurationField("azureApiVersion", "")
-								}
-							}}>
-							{t("settings:modelInfo.azureApiVersion")}
-						</Checkbox>
-						{azureApiVersionSelected && (
-							<VSCodeTextField
-								value={apiConfiguration?.azureApiVersion || ""}
-								onInput={handleInputChange("azureApiVersion")}
-								placeholder={`Default: ${azureOpenAiDefaultApiVersion}`}
-								className="w-full mt-1"
-							/>
-						)}
-					</div>
-
-					{/* Custom Headers UI */}
-					<div className="mb-4">
-						<div className="flex justify-between items-center mb-2">
-							<label className="block font-medium">{t("settings:providers.customHeaders")}</label>
-							<VSCodeButton
-								appearance="icon"
-								title={t("settings:common.add")}
-								onClick={handleAddCustomHeader}>
-								<span className="codicon codicon-add"></span>
-							</VSCodeButton>
-						</div>
-						{!customHeaders.length ? (
-							<div className="text-sm text-vscode-descriptionForeground">
-								{t("settings:providers.noCustomHeaders")}
-							</div>
-						) : (
-							customHeaders.map(([key, value], index) => (
-								<div key={index} className="flex items-center mb-2">
-									<VSCodeTextField
-										value={key}
-										className="flex-1 mr-2"
-										placeholder={t("settings:providers.headerName")}
-										onInput={(e: any) => handleUpdateHeaderKey(index, e.target.value)}
-									/>
-									<VSCodeTextField
-										value={value}
-										className="flex-1 mr-2"
-										placeholder={t("settings:providers.headerValue")}
-										onInput={(e: any) => handleUpdateHeaderValue(index, e.target.value)}
-									/>
-									<VSCodeButton
-										appearance="icon"
-										title={t("settings:common.remove")}
-										onClick={() => handleRemoveCustomHeader(index)}>
-										<span className="codicon codicon-trash"></span>
-									</VSCodeButton>
-								</div>
-							))
-						)}
-					</div>
-
-					<div className="flex flex-col gap-1">
-						<Checkbox
-							checked={apiConfiguration.enableReasoningEffort ?? false}
-							onChange={(checked: boolean) => {
-								setApiConfigurationField("enableReasoningEffort", checked)
-
-								if (!checked) {
-									const { reasoningEffort: _, ...openAiCustomModelInfo } =
-										apiConfiguration.openAiCustomModelInfo || openAiModelInfoSaneDefaults
-
-									setApiConfigurationField("openAiCustomModelInfo", openAiCustomModelInfo)
-								}
-							}}>
-							{t("settings:providers.setReasoningLevel")}
-						</Checkbox>
-						{!!apiConfiguration.enableReasoningEffort && (
-							<ReasoningEffort
-								apiConfiguration={{
-									...apiConfiguration,
-									reasoningEffort: apiConfiguration.openAiCustomModelInfo?.reasoningEffort,
-								}}
-								setApiConfigurationField={(field, value) => {
-									if (field === "reasoningEffort") {
-										const openAiCustomModelInfo =
-											apiConfiguration.openAiCustomModelInfo || openAiModelInfoSaneDefaults
-
-										setApiConfigurationField("openAiCustomModelInfo", {
-											...openAiCustomModelInfo,
-											reasoningEffort: value as ReasoningEffortType,
-										})
-									}
-								}}
-							/>
-						)}
-					</div>
-					<div className="flex flex-col gap-3">
-						<div className="text-sm text-vscode-descriptionForeground whitespace-pre-line">
-							{t("settings:providers.customModel.capabilities")}
-						</div>
-
-						<div>
-							<VSCodeTextField
-								value={
-									apiConfiguration?.openAiCustomModelInfo?.maxTokens?.toString() ||
-									openAiModelInfoSaneDefaults.maxTokens?.toString() ||
-									""
-								}
-								type="text"
-								style={{
-									borderColor: (() => {
-										const value = apiConfiguration?.openAiCustomModelInfo?.maxTokens
-
-										if (!value) {
-											return "var(--vscode-input-border)"
-										}
-
-										return value > 0
-											? "var(--vscode-charts-green)"
-											: "var(--vscode-errorForeground)"
-									})(),
-								}}
-								title={t("settings:providers.customModel.maxTokens.description")}
-								onInput={handleInputChange("openAiCustomModelInfo", (e) => {
-									const value = parseInt((e.target as HTMLInputElement).value)
-
-									return {
-										...(apiConfiguration?.openAiCustomModelInfo || openAiModelInfoSaneDefaults),
-										maxTokens: isNaN(value) ? undefined : value,
-									}
-								})}
-								placeholder={t("settings:placeholders.numbers.maxTokens")}
-								className="w-full">
-								<label className="block font-medium mb-1">
-									{t("settings:providers.customModel.maxTokens.label")}
-								</label>
-							</VSCodeTextField>
-							<div className="text-sm text-vscode-descriptionForeground">
-								{t("settings:providers.customModel.maxTokens.description")}
-							</div>
-						</div>
-
-						<div>
-							<VSCodeTextField
-								value={
-									apiConfiguration?.openAiCustomModelInfo?.contextWindow?.toString() ||
-									openAiModelInfoSaneDefaults.contextWindow?.toString() ||
-									""
-								}
-								type="text"
-								style={{
-									borderColor: (() => {
-										const value = apiConfiguration?.openAiCustomModelInfo?.contextWindow
-
-										if (!value) {
-											return "var(--vscode-input-border)"
-										}
-
-										return value > 0
-											? "var(--vscode-charts-green)"
-											: "var(--vscode-errorForeground)"
-									})(),
-								}}
-								title={t("settings:providers.customModel.contextWindow.description")}
-								onInput={handleInputChange("openAiCustomModelInfo", (e) => {
-									const value = (e.target as HTMLInputElement).value
-									const parsed = parseInt(value)
-
-									return {
-										...(apiConfiguration?.openAiCustomModelInfo || openAiModelInfoSaneDefaults),
-										contextWindow: isNaN(parsed)
-											? openAiModelInfoSaneDefaults.contextWindow
-											: parsed,
-									}
-								})}
-								placeholder={t("settings:placeholders.numbers.contextWindow")}
-								className="w-full">
-								<label className="block font-medium mb-1">
-									{t("settings:providers.customModel.contextWindow.label")}
-								</label>
-							</VSCodeTextField>
-							<div className="text-sm text-vscode-descriptionForeground">
-								{t("settings:providers.customModel.contextWindow.description")}
-							</div>
-						</div>
-
-						<div>
-							<div className="flex items-center gap-1">
-								<Checkbox
-									checked={
-										apiConfiguration?.openAiCustomModelInfo?.supportsImages ??
-										openAiModelInfoSaneDefaults.supportsImages
-									}
-									onChange={handleInputChange("openAiCustomModelInfo", (checked) => {
-										return {
-											...(apiConfiguration?.openAiCustomModelInfo || openAiModelInfoSaneDefaults),
-											supportsImages: checked,
-										}
-									})}>
-									<span className="font-medium">
-										{t("settings:providers.customModel.imageSupport.label")}
-									</span>
-								</Checkbox>
-								<i
-									className="codicon codicon-info text-vscode-descriptionForeground"
-									title={t("settings:providers.customModel.imageSupport.description")}
-									style={{ fontSize: "12px" }}
-								/>
-							</div>
-							<div className="text-sm text-vscode-descriptionForeground pt-1">
-								{t("settings:providers.customModel.imageSupport.description")}
-							</div>
-						</div>
-
-						<div>
-							<div className="flex items-center gap-1">
-								<Checkbox
-									checked={apiConfiguration?.openAiCustomModelInfo?.supportsComputerUse ?? false}
-									onChange={handleInputChange("openAiCustomModelInfo", (checked) => {
-										return {
-											...(apiConfiguration?.openAiCustomModelInfo || openAiModelInfoSaneDefaults),
-											supportsComputerUse: checked,
-										}
-									})}>
-									<span className="font-medium">
-										{t("settings:providers.customModel.computerUse.label")}
-									</span>
-								</Checkbox>
-								<i
-									className="codicon codicon-info text-vscode-descriptionForeground"
-									title={t("settings:providers.customModel.computerUse.description")}
-									style={{ fontSize: "12px" }}
-								/>
-							</div>
-							<div className="text-sm text-vscode-descriptionForeground pt-1">
-								{t("settings:providers.customModel.computerUse.description")}
-							</div>
-						</div>
-
-						<div>
-							<div className="flex items-center gap-1">
-								<Checkbox
-									checked={apiConfiguration?.openAiCustomModelInfo?.supportsPromptCache ?? false}
-									onChange={handleInputChange("openAiCustomModelInfo", (checked) => {
-										return {
-											...(apiConfiguration?.openAiCustomModelInfo || openAiModelInfoSaneDefaults),
-											supportsPromptCache: checked,
-										}
-									})}>
-									<span className="font-medium">
-										{t("settings:providers.customModel.promptCache.label")}
-									</span>
-								</Checkbox>
-								<i
-									className="codicon codicon-info text-vscode-descriptionForeground"
-									title={t("settings:providers.customModel.promptCache.description")}
-									style={{ fontSize: "12px" }}
-								/>
-							</div>
-							<div className="text-sm text-vscode-descriptionForeground pt-1">
-								{t("settings:providers.customModel.promptCache.description")}
-							</div>
-						</div>
-
-						<div>
-							<VSCodeTextField
-								value={
-									apiConfiguration?.openAiCustomModelInfo?.inputPrice?.toString() ??
-									openAiModelInfoSaneDefaults.inputPrice?.toString() ??
-									""
-								}
-								type="text"
-								style={{
-									borderColor: (() => {
-										const value = apiConfiguration?.openAiCustomModelInfo?.inputPrice
-
-										if (!value && value !== 0) {
-											return "var(--vscode-input-border)"
-										}
-
-										return value >= 0
-											? "var(--vscode-charts-green)"
-											: "var(--vscode-errorForeground)"
-									})(),
-								}}
-								onChange={handleInputChange("openAiCustomModelInfo", (e) => {
-									const value = (e.target as HTMLInputElement).value
-									const parsed = parseFloat(value)
-
-									return {
-										...(apiConfiguration?.openAiCustomModelInfo ?? openAiModelInfoSaneDefaults),
-										inputPrice: isNaN(parsed) ? openAiModelInfoSaneDefaults.inputPrice : parsed,
-									}
-								})}
-								placeholder={t("settings:placeholders.numbers.inputPrice")}
-								className="w-full">
-								<div className="flex items-center gap-1">
-									<label className="block font-medium mb-1">
-										{t("settings:providers.customModel.pricing.input.label")}
-									</label>
-									<i
-										className="codicon codicon-info text-vscode-descriptionForeground"
-										title={t("settings:providers.customModel.pricing.input.description")}
-										style={{ fontSize: "12px" }}
-									/>
-								</div>
-							</VSCodeTextField>
-						</div>
-
-						<div>
-							<VSCodeTextField
-								value={
-									apiConfiguration?.openAiCustomModelInfo?.outputPrice?.toString() ||
-									openAiModelInfoSaneDefaults.outputPrice?.toString() ||
-									""
-								}
-								type="text"
-								style={{
-									borderColor: (() => {
-										const value = apiConfiguration?.openAiCustomModelInfo?.outputPrice
-
-										if (!value && value !== 0) {
-											return "var(--vscode-input-border)"
-										}
-
-										return value >= 0
-											? "var(--vscode-charts-green)"
-											: "var(--vscode-errorForeground)"
-									})(),
-								}}
-								onChange={handleInputChange("openAiCustomModelInfo", (e) => {
-									const value = (e.target as HTMLInputElement).value
-									const parsed = parseFloat(value)
-
-									return {
-										...(apiConfiguration?.openAiCustomModelInfo || openAiModelInfoSaneDefaults),
-										outputPrice: isNaN(parsed) ? openAiModelInfoSaneDefaults.outputPrice : parsed,
-									}
-								})}
-								placeholder={t("settings:placeholders.numbers.outputPrice")}
-								className="w-full">
-								<div className="flex items-center gap-1">
-									<label className="block font-medium mb-1">
-										{t("settings:providers.customModel.pricing.output.label")}
-									</label>
-									<i
-										className="codicon codicon-info text-vscode-descriptionForeground"
-										title={t("settings:providers.customModel.pricing.output.description")}
-										style={{ fontSize: "12px" }}
-									/>
-								</div>
-							</VSCodeTextField>
-						</div>
-
-						{apiConfiguration?.openAiCustomModelInfo?.supportsPromptCache && (
-							<>
-								<div>
-									<VSCodeTextField
-										value={
-											apiConfiguration?.openAiCustomModelInfo?.cacheReadsPrice?.toString() ?? "0"
-										}
-										type="text"
-										style={{
-											borderColor: (() => {
-												const value = apiConfiguration?.openAiCustomModelInfo?.cacheReadsPrice
-
-												if (!value && value !== 0) {
-													return "var(--vscode-input-border)"
-												}
-
-												return value >= 0
-													? "var(--vscode-charts-green)"
-													: "var(--vscode-errorForeground)"
-											})(),
-										}}
-										onChange={handleInputChange("openAiCustomModelInfo", (e) => {
-											const value = (e.target as HTMLInputElement).value
-											const parsed = parseFloat(value)
-
-											return {
-												...(apiConfiguration?.openAiCustomModelInfo ??
-													openAiModelInfoSaneDefaults),
-												cacheReadsPrice: isNaN(parsed) ? 0 : parsed,
-											}
-										})}
-										placeholder={t("settings:placeholders.numbers.inputPrice")}
-										className="w-full">
-										<div className="flex items-center gap-1">
-											<span className="font-medium">
-												{t("settings:providers.customModel.pricing.cacheReads.label")}
-											</span>
-											<i
-												className="codicon codicon-info text-vscode-descriptionForeground"
-												title={t(
-													"settings:providers.customModel.pricing.cacheReads.description",
-												)}
-												style={{ fontSize: "12px" }}
-											/>
-										</div>
-									</VSCodeTextField>
-								</div>
-								<div>
-									<VSCodeTextField
-										value={
-											apiConfiguration?.openAiCustomModelInfo?.cacheWritesPrice?.toString() ?? "0"
-										}
-										type="text"
-										style={{
-											borderColor: (() => {
-												const value = apiConfiguration?.openAiCustomModelInfo?.cacheWritesPrice
-
-												if (!value && value !== 0) {
-													return "var(--vscode-input-border)"
-												}
-
-												return value >= 0
-													? "var(--vscode-charts-green)"
-													: "var(--vscode-errorForeground)"
-											})(),
-										}}
-										onChange={handleInputChange("openAiCustomModelInfo", (e) => {
-											const value = (e.target as HTMLInputElement).value
-											const parsed = parseFloat(value)
-
-											return {
-												...(apiConfiguration?.openAiCustomModelInfo ??
-													openAiModelInfoSaneDefaults),
-												cacheWritesPrice: isNaN(parsed) ? 0 : parsed,
-											}
-										})}
-										placeholder={t("settings:placeholders.numbers.cacheWritePrice")}
-										className="w-full">
-										<div className="flex items-center gap-1">
-											<label className="block font-medium mb-1">
-												{t("settings:providers.customModel.pricing.cacheWrites.label")}
-											</label>
-											<i
-												className="codicon codicon-info text-vscode-descriptionForeground"
-												title={t(
-													"settings:providers.customModel.pricing.cacheWrites.description",
-												)}
-												style={{ fontSize: "12px" }}
-											/>
-										</div>
-									</VSCodeTextField>
-								</div>
-							</>
-						)}
-
-						<Button
-							variant="secondary"
-							onClick={() =>
-								setApiConfigurationField("openAiCustomModelInfo", openAiModelInfoSaneDefaults)
-							}>
-							{t("settings:providers.customModel.resetDefaults")}
-						</Button>
-					</div>
-				</>
+				<OpenAICompatible
+					apiConfiguration={apiConfiguration}
+					setApiConfigurationField={setApiConfigurationField}
+				/>
 			)}
 
 			{selectedProvider === "lmstudio" && (
-				<>
-					<VSCodeTextField
-						value={apiConfiguration?.lmStudioBaseUrl || ""}
-						type="url"
-						onInput={handleInputChange("lmStudioBaseUrl")}
-						placeholder={t("settings:defaults.lmStudioUrl")}
-						className="w-full">
-						<label className="block font-medium mb-1">{t("settings:providers.lmStudio.baseUrl")}</label>
-					</VSCodeTextField>
-					<VSCodeTextField
-						value={apiConfiguration?.lmStudioModelId || ""}
-						onInput={handleInputChange("lmStudioModelId")}
-						placeholder={t("settings:placeholders.modelId.lmStudio")}
-						className="w-full">
-						<label className="block font-medium mb-1">{t("settings:providers.lmStudio.modelId")}</label>
-					</VSCodeTextField>
-					{lmStudioModels.length > 0 && (
-						<VSCodeRadioGroup
-							value={
-								lmStudioModels.includes(apiConfiguration?.lmStudioModelId || "")
-									? apiConfiguration?.lmStudioModelId
-									: ""
-							}
-							onChange={handleInputChange("lmStudioModelId")}>
-							{lmStudioModels.map((model) => (
-								<VSCodeRadio
-									key={model}
-									value={model}
-									checked={apiConfiguration?.lmStudioModelId === model}>
-									{model}
-								</VSCodeRadio>
-							))}
-						</VSCodeRadioGroup>
-					)}
-					<Checkbox
-						checked={apiConfiguration?.lmStudioSpeculativeDecodingEnabled === true}
-						onChange={(checked) => {
-							setApiConfigurationField("lmStudioSpeculativeDecodingEnabled", checked)
-						}}>
-						{t("settings:providers.lmStudio.speculativeDecoding")}
-					</Checkbox>
-					{apiConfiguration?.lmStudioSpeculativeDecodingEnabled && (
-						<>
-							<div>
-								<VSCodeTextField
-									value={apiConfiguration?.lmStudioDraftModelId || ""}
-									onInput={handleInputChange("lmStudioDraftModelId")}
-									placeholder={t("settings:placeholders.modelId.lmStudioDraft")}
-									className="w-full">
-									<label className="block font-medium mb-1">
-										{t("settings:providers.lmStudio.draftModelId")}
-									</label>
-								</VSCodeTextField>
-								<div className="text-sm text-vscode-descriptionForeground">
-									{t("settings:providers.lmStudio.draftModelDesc")}
-								</div>
-							</div>
-							{lmStudioModels.length > 0 && (
-								<>
-									<div className="font-medium">
-										{t("settings:providers.lmStudio.selectDraftModel")}
-									</div>
-									<VSCodeRadioGroup
-										value={
-											lmStudioModels.includes(apiConfiguration?.lmStudioDraftModelId || "")
-												? apiConfiguration?.lmStudioDraftModelId
-												: ""
-										}
-										onChange={handleInputChange("lmStudioDraftModelId")}>
-										{lmStudioModels.map((model) => (
-											<VSCodeRadio key={`draft-${model}`} value={model}>
-												{model}
-											</VSCodeRadio>
-										))}
-									</VSCodeRadioGroup>
-									{lmStudioModels.length === 0 && (
-										<div
-											className="text-sm rounded-xs p-2"
-											style={{
-												backgroundColor: "var(--vscode-inputValidation-infoBackground)",
-												border: "1px solid var(--vscode-inputValidation-infoBorder)",
-												color: "var(--vscode-inputValidation-infoForeground)",
-											}}>
-											{t("settings:providers.lmStudio.noModelsFound")}
-										</div>
-									)}
-								</>
-							)}
-						</>
-					)}
-					<div className="text-sm text-vscode-descriptionForeground">
-						<Trans
-							i18nKey="settings:providers.lmStudio.description"
-							components={{
-								a: <VSCodeLink href="https://lmstudio.ai/docs" />,
-								b: <VSCodeLink href="https://lmstudio.ai/docs/basics/server" />,
-								span: (
-									<span className="text-vscode-errorForeground ml-1">
-										<span className="font-medium">Note:</span>
-									</span>
-								),
-							}}
-						/>
-					</div>
-				</>
+				<LMStudio apiConfiguration={apiConfiguration} setApiConfigurationField={setApiConfigurationField} />
 			)}
 
 			{selectedProvider === "deepseek" && (
@@ -1511,85 +475,11 @@ const ApiOptions = ({
 			)}
 
 			{selectedProvider === "vscode-lm" && (
-				<>
-					<div>
-						<label className="block font-medium mb-1">{t("settings:providers.vscodeLmModel")}</label>
-						{vsCodeLmModels.length > 0 ? (
-							<Select
-								value={
-									apiConfiguration?.vsCodeLmModelSelector
-										? `${apiConfiguration.vsCodeLmModelSelector.vendor ?? ""}/${apiConfiguration.vsCodeLmModelSelector.family ?? ""}`
-										: ""
-								}
-								onValueChange={handleInputChange("vsCodeLmModelSelector", (value) => {
-									const [vendor, family] = value.split("/")
-									return { vendor, family }
-								})}>
-								<SelectTrigger className="w-full">
-									<SelectValue placeholder={t("settings:common.select")} />
-								</SelectTrigger>
-								<SelectContent>
-									{vsCodeLmModels.map((model) => (
-										<SelectItem
-											key={`${model.vendor}/${model.family}`}
-											value={`${model.vendor}/${model.family}`}>
-											{`${model.vendor} - ${model.family}`}
-										</SelectItem>
-									))}
-								</SelectContent>
-							</Select>
-						) : (
-							<div className="text-sm text-vscode-descriptionForeground">
-								{t("settings:providers.vscodeLmDescription")}
-							</div>
-						)}
-					</div>
-					<div className="text-sm text-vscode-errorForeground">{t("settings:providers.vscodeLmWarning")}</div>
-				</>
+				<VSCodeLM apiConfiguration={apiConfiguration} setApiConfigurationField={setApiConfigurationField} />
 			)}
 
 			{selectedProvider === "ollama" && (
-				<>
-					<VSCodeTextField
-						value={apiConfiguration?.ollamaBaseUrl || ""}
-						type="url"
-						onInput={handleInputChange("ollamaBaseUrl")}
-						placeholder={t("settings:defaults.ollamaUrl")}
-						className="w-full">
-						<label className="block font-medium mb-1">{t("settings:providers.ollama.baseUrl")}</label>
-					</VSCodeTextField>
-					<VSCodeTextField
-						value={apiConfiguration?.ollamaModelId || ""}
-						onInput={handleInputChange("ollamaModelId")}
-						placeholder={t("settings:placeholders.modelId.ollama")}
-						className="w-full">
-						<label className="block font-medium mb-1">{t("settings:providers.ollama.modelId")}</label>
-					</VSCodeTextField>
-					{ollamaModels.length > 0 && (
-						<VSCodeRadioGroup
-							value={
-								ollamaModels.includes(apiConfiguration?.ollamaModelId || "")
-									? apiConfiguration?.ollamaModelId
-									: ""
-							}
-							onChange={handleInputChange("ollamaModelId")}>
-							{ollamaModels.map((model) => (
-								<VSCodeRadio
-									key={model}
-									value={model}
-									checked={apiConfiguration?.ollamaModelId === model}>
-									{model}
-								</VSCodeRadio>
-							))}
-						</VSCodeRadioGroup>
-					)}
-					<div className="text-sm text-vscode-descriptionForeground">
-						{t("settings:providers.ollama.description")}
-						<span className="text-vscode-errorForeground ml-1">
-							{t("settings:providers.ollama.warning")}
-						</span>
-					</div>
-				</>
+				<Ollama apiConfiguration={apiConfiguration} setApiConfigurationField={setApiConfigurationField} />
 			)}
 
 			{selectedProvider === "xai" && (
@@ -1647,12 +537,11 @@ const ApiOptions = ({
 					<div className="text-sm text-vscode-descriptionForeground -mt-2">
 						{t("settings:providers.apiKeyStorageNotice")}
 					</div>
-					{/* Add a link to get Chutes API key if available */}
-					{/* {!apiConfiguration?.chutesApiKey && (
-						<VSCodeButtonLink href="LINK_TO_CHUTES_API_KEYS" appearance="secondary">
+					{!apiConfiguration?.chutesApiKey && (
+						<VSCodeButtonLink href="https://chutes.ai/app/api" appearance="secondary">
 							{t("settings:providers.getChutesApiKey")}
 						</VSCodeButtonLink>
-					)} */}
+					)}
 				</>
 			)}
