+++ webview-ui/src/components/settings/transforms.ts
@@ -1,0 +1,3 @@
+export const noTransform = <T>(value: T) => value
+
+export const inputEventTransform = <E>(event: E) => (event as { target: HTMLInputElement })?.target?.value as any
